requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
cryptography
playwright
